+++ CHANGELOG.md
@@ -19,9 +19,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - A way to export settings as a way to create and share presets with other worlds and users.
 
 ## [0.1.1] - 2021-02-02
+### Added
+- Colour Picker to change the status checkbox colours. (The checkmark currently cannot be edited.)
 ### Fixed
 - Item and Vehicle Sheets are now affected.
-- Colour Picker to change the status checkbox colours. (The checkmark currently cannot be edited.)
 
 ## [0.1.0] - 2021-02-01
 ### Added